Feature: libcrmcommon: crm_ipc_send can receive large IPC messages.
This function is used to send regular IPC messages from a client to a
server.  However, it is also used to receive synchronous replies to
those messages.  Thus, it has a helper function for receiving.

Just like with crm_ipc_read, we will read data into a temporary buffer
and append it to the client's read buffer until we've either read all
parts of the message or read the entirety of a single message.

This function is still subject to an overall timeout, as well as looping
to try reading again in the event that we get EAGAIN from libqb.