Feature: libcrmcommon: pcmk__ipc_send_xml can send large IPC messages.
These function names are a little hard to follow, but this is actually
the companion code to the receive block in crm_ipc_send.  In this case,
pcmk__ipc_send_xml is sending regular IPC messages as responses to some
client request.  Those requests are sent from crm_ipc_send, which is
waiting for a synchronous reply.

We do this by basically following the same approach as we did for server
events in pcmk__ipc_send_xml.  The only difference is that we need real
EAGAIN handling because there's no send queue for IPC messages.