Low: cts: Fix the BadNews messages that the SplitBrain test ignores.
This has been broken since the message changed in 07a19bcf3be, but we
didn't notice because the SplitBrain test has been disabled for a long
time.

Additionally, I've removed the "stonith " part from the beginning of the
string.  Other resources can cause the same message to be printed in the
SplitBrain test.  We could have a whole separate message for this case,
but I'm not sure it's worth it.